changed det to double in main and added const in .h

+++ src/main.cpp
@@ -5,20 +5,20 @@ int main() {
 	size_t n = 0;
 	std::cin >> n;
 
-	std::vector <std::vector<int>> vec;
+	std::vector <std::vector<double>> vec;
 	for (int i = 0; i < n; i ++ ) {
-		std::vector<int>vec_2;
+		std::vector<double>vec_2;
 		vec.push_back(vec_2);
 	}
 	for (int i = 0; i < n; ++i) {
 		for (int j = 0; j < n; ++j) {
-			int elem;
+			double elem;
 			std::cin >> elem;
-			std::vector<int> &pr_vec = vec[i];
+			std::vector<double> &pr_vec = vec[i];
 			pr_vec.push_back(elem);
 		}
 	}
-	Matrix<int> mat(n, vec);
-	std::cout << static_cast<long long>(mat.det()) << "\n";
+	Matrix<double> mat(n, vec);
+	std::cout << mat.det() << "\n";
 	return 0;
 }
